This diff enables module API loading program memory-planed on non-cpu device. It update Module::load_method() to detect device buffers via MethodMeta and allocate device memory using the registered DeviceAllocator. Device memory is managed via DeviceMemoryBuffer RAII objects stored in PlannedMemory, ensuring proper cleanup when the Method is destroyed.

Differential Revision: [D97850705](https://our.internmc.facebook.com/intern/diff/D97850705/)

[ghstack-poisoned]

+++ examples/models/llama/tests/test_static_attention.py
@@ -325,7 +325,7 @@ class StaticAttentionTest(unittest.TestCase):
             static_config, input_len, cache_len, batch_size=batch_size
         )
         example_inputs = (
-            torch.zeros(batch_size, input_len),
+            torch.zeros(batch_size, input_len, dtype=torch.long),
             {
                 "masks": mgr.masks,
                 "freqs_cos_override": mgr.freqs_cos[:input_len],
@@ -350,7 +350,7 @@ class StaticAttentionTest(unittest.TestCase):
             static_config, input_len, cache_len, batch_size=1
         )
         example_inputs = (
-            torch.zeros(1, input_len),
+            torch.zeros(1, input_len, dtype=torch.long),
             {
                 "masks": mgr.masks,
                 "freqs_cos_override": mgr.freqs_cos[:input_len],
